refactor

+++ libft/ft_strlen.c
@@ -9,11 +9,11 @@
 /*   Updated: 2025/04/17 17:42:28 by andcardo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
-#include <stddef.h>
+#include "libft.h"
 
 size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
